Split RunHelper and simplify the Execute method for both cases.
